8.4.3 Derivation process for prediction weights

+++ src/video/atom/avcc/sps.rs
@@ -124,18 +124,18 @@ impl SequenceParameterSet {
 #[derive(Debug)]
 pub struct PicOrderCntTypeOne {
   pub delta_pic_order_always_zero_flag: bool,
-  pub offset_for_non_ref_pic: i16,
-  pub offset_for_top_to_bottom_field: i16,
+  pub offset_for_non_ref_pic: isize,
+  pub offset_for_top_to_bottom_field: isize,
   pub num_ref_frames_in_pic_order_cnt_cycle: u16,
-  pub offset_for_ref_frame: Box<[i16]>,
-  pub expected_delta_per_pic_order_cnt_cycle: i16,
+  pub offset_for_ref_frame: Box<[isize]>,
+  pub expected_delta_per_pic_order_cnt_cycle: isize,
 }
 
 impl PicOrderCntTypeOne {
   pub fn new(data: &mut BitStream, pic_order_cnt_type: u16) -> Option<Self> {
     (pic_order_cnt_type == 1).then(|| {
       let num_ref_frames_in_pic_order_cnt_cycle;
-      let offset_for_ref_frame: Box<[i16]>;
+      let offset_for_ref_frame: Box<[isize]>;
       let expected_delta_per_pic_order_cnt_cycle;
       Self {
         delta_pic_order_always_zero_flag: data.bit_flag(),
